@@ -14,7 +14,7 @@ description: A virtual secretary
 version: 1.1.2
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
