@@ -1,9 +1,10 @@
+import 'package:auto_call/classes/setting.dart';
 import 'package:flutter/material.dart';
 
 import 'package:permission_handler/permission_handler.dart';
 import 'package:contacts_service/contacts_service.dart';
 
-import 'package:auto_call/services/settings_manager.dart';
+import 'package:auto_call/classes/settings_manager.dart';
 import 'package:auto_call/pages/upgrade.dart';
 import 'package:auto_call/ui/widgets/permission_widget.dart';
 import 'package:auto_call/ui/widgets/phone_contacts.dart';
