@@ -6,7 +6,7 @@ class DonutPieChart extends StatelessWidget {
   final List<charts.Series> seriesList;
   final bool animate;
 
-  DonutPieChart(this.seriesList, {this.animate});
+  DonutPieChart(this.seriesList, {required this.animate});
 
   /// Creates a [PieChart] with sample data and no transition.
   factory DonutPieChart.withSampleData() {
@@ -17,17 +17,15 @@ class DonutPieChart extends StatelessWidget {
     );
   }
 
-
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: EdgeInsets.all(10),
+        padding: EdgeInsets.all(10),
         child: charts.PieChart(seriesList,
-        animate: animate,
-        // Configure the width of the pie slices to 60px. The remaining space in
-        // the chart will be left as a hole in the center.
-        defaultRenderer: new charts.ArcRendererConfig(arcWidth: 60))
-    );
+            animate: animate,
+            // Configure the width of the pie slices to 60px. The remaining space in
+            // the chart will be left as a hole in the center.
+            defaultRenderer: new charts.ArcRendererConfig(arcWidth: 60)));
   }
 
   /// Create one series with sample hard coded data.
